fix: correção function para exibir cotações

+++ src/Digitas.TesteTecnico/Digitas.QuotesMetrics.Function/Domain/Entities/EthAsk.cs
@@ -4,6 +4,6 @@ public class EthAsk
 {
     public int EthAskId { get; set; }
     public long Microtimestamp { get; set; }
-    public int UsdValue { get; set; }
+    public decimal UsdValue { get; set; }
     public decimal Amount { get; set; }
 }
