+++ src/Digitas.TesteTecnico/Digitas.QuotesMetrics.Function/Domain/Entities/EthBid.cs
@@ -4,6 +4,6 @@ public class EthBid
 {
     public int EthBidId { get; set; }
     public long Microtimestamp { get; set; }
-    public int UsdValue { get; set; }
+    public decimal UsdValue { get; set; }
     public decimal Amount { get; set; }
 }
