+++ src/Digitas.TesteTecnico/Digitas.QuotesMetrics.Function/Domain/ValueObject/Metrics.cs
@@ -2,8 +2,8 @@
 
 public class Metrics
 {
-    public int BiggestUsdValue { get; set; }
-    public int SmallestUsdValue { get; set; }
+    public decimal BiggestUsdValue { get; set; }
+    public decimal SmallestUsdValue { get; set; }
     public decimal AvgUsdValue { get; set; }
     public decimal AvgLastFiveMinutesUsdValue { get; set; }
     public decimal AvgAmount { get; set; }
